0704 수업전복습

+++ src/section03/Practice0307.java
@@ -1,10 +1,10 @@
 package section03;
+//check
 public class Practice0307 {
     public static void main(String[] args) {
       //형변환 test
-      double a = 5;
-      int b = 2;
-
+      int a = 5;
+      double b = 2;
 
       System.out.println("출력 결과 2.5\t: " + (a / b));
